Gather all std::op impls for `ApInt`
