+++ src/errors.rs
@@ -126,7 +126,7 @@ pub enum DivOp {
     SignedRem,
 }
 
-/// Represents an error that may occure upon using the `ApInt` library.
+/// Represents an error that may occur upon using the `ApInt` library.
 ///
 /// All errors have a unique kind which also stores extra information for error
 /// reporting. Besides that an `Error` also stores a message and an optional
